requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
